🥪 ss2022/tcp: change padding behavior
Previously, for Shadowsocks 2022 TCP, padding is only added when there's no initial payload. This commit changes the behavior so that padding is added as long as the size of the initial payload does not exceed MaxPaddingLength.

Along with the above change is a bit of refactoring and a fix for large (>128k) initial payload handling.