🥧 httpproxy: prevent goroutine leak from failed excessive pipelining
If the client sent an excessive amount of pipelined requests, and the server closed the connection before sending out all the responses, the request forwarding goroutine would be stuck sending to reqCh.

To unblock the goroutine, this commit adds another channel that gets closed when the remote connection is closed.